requirements.txt
transformers
torch
torchvision
opencv-python
Pillow
mediapipe
flask
flask-cors
numpy